requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
nanoid==2.0.0 